refactor(pkg/abac/pdp): refactor pdp eval, part 1
