+++ pkg/abac/pdp/translate/translate.go
@@ -14,17 +14,15 @@ import (
 	"errors"
 	"fmt"
 
-	"iam/pkg/abac/pdp/types"
-	"iam/pkg/abac/pdp/util"
+	jsoniter "github.com/json-iterator/go"
+
+	"iam/pkg/abac/pdp/condition"
+	pdptypes "iam/pkg/abac/pdp/types"
+	"iam/pkg/errorx"
 )
 
-/*
-表达式转换
-*/
-
-// 1. 解析policy表达式到对外的条件表达式
-// 2. 对操作符需要有一个映射关系
-// 3. 遍历实现
+// Translate ...
+const Translate = "Translate"
 
 var errMustNotEmpty = errors.New("value must not be empty")
 
@@ -36,163 +34,191 @@ func (c ExprCell) Op() string {
 	return c["op"].(string)
 }
 
-type translateFunc func(field string, value []interface{}) (ExprCell, error)
+// PoliciesTranslate 策略列表转换为QL表达式
+func PoliciesTranslate(
+	policies []condition.Condition,
+) (map[string]interface{}, error) {
+	errorWrapf := errorx.NewLayerFunctionErrorWrapf(Translate, "PoliciesTranslate")
 
-var translateFactories map[string]translateFunc
-
-func init() {
-	translateFactories = map[string]translateFunc{
-		"AND":           andTranslate,
-		"OR":            orTranslate,
-		"Any":           anyTranslate,
-		"StringEquals":  stringEqualsTranslate,
-		"StringPrefix":  stringPrefixTranslate,
-		"NumericEquals": numericEqualsTranslate,
-		"Bool":          boolTranslate,
-	}
-}
-
-func singleTranslate(expression types.PolicyCondition, _type string) (ExprCell, error) {
-	for operator, option := range expression {
-		tf, ok := translateFactories[operator]
-		if !ok {
-			return nil, fmt.Errorf("can not support operator %s", operator)
-		}
-
-		for field, value := range option {
-			switch operator {
-			case "OR", "AND":
-				return tf(_type, value)
-			default:
-				//typeField := fmt.Sprintf("%s.%s", _type, field)
-				typeField := _type + "." + field
-				return tf(typeField, value)
-			}
-		}
-	}
-	return nil, errMustNotEmpty
-}
-
-func andTranslate(_type string, value []interface{}) (ExprCell, error) {
-	content := make([]interface{}, 0, len(value))
-
-	for _, v := range value {
-		m, err := util.InterfaceToPolicyCondition(v)
+	content := make([]ExprCell, 0, len(policies))
+	for _, policy := range policies {
+		// TODO: 可以优化的点, expression + resourceTypeSet => Condition的local cache
+		condition, err := policyTranslate(policy)
 		if err != nil {
+			err = errorWrapf(err, "policyTranslate condition=`%+v` fail", policy)
 			return nil, err
 		}
-		condition, err2 := singleTranslate(m, _type)
-		if err2 != nil {
-			return nil, err2
+
+		// NOTE: if got an `any`, return `any`!
+		if condition.Op() == "any" {
+			return ExprCell{
+				"op":    "any",
+				"field": "",
+				"value": []string{},
+			}, nil
 		}
 
 		content = append(content, condition)
+
 	}
 
-	return map[string]interface{}{
-		"op":      "AND",
-		"content": content,
-	}, nil
-}
-
-func orTranslate(_type string, value []interface{}) (ExprCell, error) {
-	content := make([]interface{}, 0, len(value))
-
-	for _, v := range value {
-		m, err := util.InterfaceToPolicyCondition(v)
-		if err != nil {
-			return nil, err
-		}
-		condition, err2 := singleTranslate(m, _type)
-		if err2 != nil {
-			return nil, err2
-		}
-
-		content = append(content, condition)
-	}
-
-	return map[string]interface{}{
-		"op":      "OR",
-		"content": content,
-	}, nil
-}
-
-//nolint:unparam
-func anyTranslate(field string, value []interface{}) (ExprCell, error) {
-	return map[string]interface{}{
-		"op":    "any",
-		"field": field,
-		"value": value,
-	}, nil
-}
-
-func stringEqualsTranslate(field string, value []interface{}) (ExprCell, error) {
-	exprCell := map[string]interface{}{
-		"field": field,
-	}
-
-	switch len(value) {
-	case 0:
-		return nil, errMustNotEmpty
-	case 1:
-		exprCell["op"] = "eq"
-		exprCell["value"] = value[0]
-	default:
-		exprCell["op"] = "in"
-		exprCell["value"] = value
-	}
-	return exprCell, nil
-}
-
-func stringPrefixTranslate(field string, value []interface{}) (ExprCell, error) {
-	content := make([]map[string]interface{}, 0, len(value))
-	for _, v := range value {
-		content = append(content, map[string]interface{}{
-			"op":    "starts_with",
-			"field": field,
-			"value": v,
-		})
+	// merge same field `eq` and `in`; to `in`
+	if len(content) > 1 {
+		// 合并条件中field相同, op为eq, in的条件
+		content = mergeContentField(content)
 	}
 
 	switch len(content) {
-	case 0:
-		return nil, errMustNotEmpty
 	case 1:
 		return content[0], nil
 	default:
-		return map[string]interface{}{
+		return ExprCell{
 			"op":      "OR",
 			"content": content,
 		}, nil
 	}
 }
 
-func numericEqualsTranslate(field string, value []interface{}) (ExprCell, error) {
-	exprCell := map[string]interface{}{
-		"field": field,
-	}
-
-	switch len(value) {
-	case 0:
-		return nil, errMustNotEmpty
-	case 1:
-		exprCell["op"] = "eq"
-		exprCell["value"] = value[0]
-	default:
-		exprCell["op"] = "in"
-		exprCell["value"] = value
-	}
-	return exprCell, nil
+func policyTranslate(
+	cond condition.Condition,
+) (ExprCell, error) {
+	return cond.Translate()
 }
 
-func boolTranslate(field string, value []interface{}) (ExprCell, error) {
-	if len(value) != 1 {
-		return nil, fmt.Errorf("bool not support multi value %+v", value)
+func PolicyStringTranslate(resourceExpression string) (ExprCell, error) {
+
+	// TODO: 代码重复
+	//pkg/cache/impls/local_unmarshaled_expression.go
+
+	expressions := []pdptypes.ResourceExpression{}
+	err := jsoniter.UnmarshalFromString(resourceExpression, &expressions)
+	// 无效的policy条件表达式, 容错
+	if err != nil {
+		err = fmt.Errorf("cache UnmarshalExpression unmarshal %s error: %w",
+			resourceExpression, err)
+		return nil, err
 	}
 
-	return map[string]interface{}{
-		"op":    "eq",
-		"field": field,
-		"value": value[0],
-	}, nil
+	content := make([]condition.Condition, 0, len(expressions))
+	for _, expression := range expressions {
+		// NOTE: change the expression
+		pc, err1 := expression.ToNewPolicyCondition()
+		if err1 != nil {
+			return nil, fmt.Errorf("toNewPolicyCondition error: %w", err1)
+		}
+
+		c, err2 := condition.NewConditionFromPolicyCondition(pc)
+		// 表达式解析出错, 容错
+		if err2 != nil {
+			return nil, fmt.Errorf("newConditionFromPolicyCondition error: %w", err2)
+		}
+		content = append(content, c)
+	}
+
+	if len(content) == 1 {
+		return content[0].Translate()
+	} else {
+		return condition.NewAndCondition(content).Translate()
+	}
+}
+
+// PolicyTranslate ...
+//func PolicyTranslate(
+//	cond condition.Condition,
+//) (ExprCell, error) {
+//	return cond.Translate()
+//errorWrapf := errorx.NewLayerFunctionErrorWrapf(Translate, "PolicyTranslate")
+
+//expressions := []pdptypes.ResourceExpression{}
+
+// TODO: newExpression, do translate here
+//       需要支持, 将新版本表达式搞过来, 支持translate
+// NOTE: if expression == "" or expression == "[]", all return any
+// if action without resource_types, the expression is ""
+//if resourceExpression != "" {
+//	err := jsoniter.UnmarshalFromString(resourceExpression, &expressions)
+//	if err != nil {
+//		err = errorWrapf(err, "unmarshal resourceExpression=`%s` fail", resourceExpression)
+//		return nil, err
+//	}
+//}
+
+// 注意, 如果resourceType不匹配, 那么最终会返回any => 这里有没有问题? 两阶段计算?
+//content := make([]ExprCell, 0, len(expressions))
+//for _, expression := range expressions {
+//	key := expression.System + ":" + expression.Type
+//	if resourceTypeSet.Has(key) {
+//		expr, err := singleTranslate(expression.Expression, expression.Type)
+//		if err != nil {
+//			err = errorWrapf(err, "pdp PolicyTranslate expression: %s", expression.Expression)
+//			return nil, err
+//		}
+//		content = append(content, expr)
+//	}
+//}
+
+//switch len(content) {
+//// content为空, 说明policy的操作不关联资源, 返回any
+//case 0:
+//	return ExprCell{
+//		"op":    "any",
+//		"field": "",
+//		"value": []string{},
+//	}, nil
+//case 1:
+//	return content[0], nil
+//default:
+//	// NOTE: 这里是满足 一个操作依赖两个资源的场景, 所以是 AND => 两阶段计算
+//	return ExprCell{
+//		"op":      "AND",
+//		"content": content,
+//	}, nil
+//}
+//}
+
+func mergeContentField(content []ExprCell) []ExprCell {
+	mergeableExprs := map[string][]ExprCell{}
+	newContent := make([]ExprCell, 0, len(content))
+
+	for _, expr := range content {
+		switch expr.Op() {
+		case "eq", "in":
+			field := expr["field"].(string)
+			exprs, ok := mergeableExprs[field]
+			if ok {
+				exprs = append(exprs, expr)
+				mergeableExprs[field] = exprs
+			} else {
+				mergeableExprs[field] = []ExprCell{expr}
+			}
+		default:
+			newContent = append(newContent, expr)
+		}
+	}
+
+	for field, exprs := range mergeableExprs {
+		if len(exprs) == 1 {
+			newContent = append(newContent, exprs[0])
+		} else {
+			values := make([]interface{}, 0, len(exprs))
+
+			// 合并
+			for _, expr := range exprs {
+				switch expr.Op() {
+				case "eq":
+					values = append(values, expr["value"])
+				case "in":
+					values = append(values, expr["value"].([]interface{})...)
+				}
+			}
+
+			newContent = append(newContent, ExprCell{
+				"op":    "in",
+				"field": field,
+				"value": values,
+			})
+		}
+	}
+
+	return newContent
 }
